+++ modules/features/taskerplugin/src/main/java/au/com/shiftyjelly/pocketcasts/taskerplugin/base/Extensions.kt
@@ -4,6 +4,8 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.platform.LocalConfiguration
 import androidx.compose.ui.unit.DpSize
 import androidx.compose.ui.unit.dp
+import java.text.SimpleDateFormat
+import java.util.*
 
 val String?.nullIfEmpty get() = if (isNullOrEmpty()) null else this
 fun <T> tryOrNull(handleError: ((Throwable) -> T?)? = null, block: () -> T?): T? = try {
@@ -21,3 +23,6 @@ val screenSize
         val screenWidth = configuration.screenWidthDp.dp
         return DpSize(width = screenWidth, height = screenHeight)
     }
+
+val Date.formattedForTasker get() = SimpleDateFormat("yyyy-MM-dd HH:mm:ss").format(this)
+val String.formattedForTasker get() = replace(",", "")