+++ modules/features/taskerplugin/src/main/java/au/com/shiftyjelly/pocketcasts/taskerplugin/queryfilterepisodes/InputQueryFilterEpisodes.kt
@@ -1,0 +1,9 @@
+package au.com.shiftyjelly.pocketcasts.taskerplugin.queryfilterepisodes
+
+import com.joaomgcd.taskerpluginlibrary.input.TaskerInputField
+import com.joaomgcd.taskerpluginlibrary.input.TaskerInputRoot
+
+@TaskerInputRoot
+class InputQueryFilterEpisodes @JvmOverloads constructor(
+    @field:TaskerInputField("title_or_id", labelResIdName = "filter_id_or_title") var titleOrId: String? = null
+)